Changes per #1521 review
